move the maybeInvertExpansion() stuff into IndexLinkOptimizer so everyone can take advantage of it.
